+++ swifttools/swift_too/query_job.py
@@ -1,19 +1,19 @@
-from .swift_clock import Swift_Clock
-from .swift_saa import Swift_SAA
-from .swift_guano import Swift_GUANO
-from .swift_calendar import Swift_Calendar
-from .api_status import Swift_TOO_Status
-from .swift_visquery import Swift_VisQuery
-from .swift_obsquery import Swift_AFST
-from .swift_planquery import Swift_PPST
-from .swift_requests import Swift_TOO_Requests
-from .swift_data import Swift_Data
-from .swift_uvot import UVOT_mode
+from .swift_clock import Clock
+from .swift_saa import SAA
+from .swift_guano import GUANO
+from .swift_calendar import Calendar
+from .api_status import TOOStatus
+from .swift_visquery import VisQuery
+from .swift_obsquery import AFST
+from .swift_planquery import PPST
+from .swift_requests import TOORequests
+from .swift_data import Data
+from .swift_uvot import UVOTMode
 
 
-class QueryJob(Swift_TOO_Status):
+class QueryJob(TOOStatus):
     """Class that enables fetching the results of already submitted jobs.
-    Essentially the same as Swift_TOO_Status other than if the process has
+    Essentially the same as TOOStatus other than if the process has
     been completed the result of the job will be attached as `result`.
 
     Attributes
@@ -37,7 +37,7 @@ class QueryJob(Swift_TOO_Status):
     warnings : list
         list of warning strings associated with the request
     too_id : list
-        For a Swift_TOO request, the TOO ID assigned to a new request
+        For a TOO request, the TOO ID assigned to a new request
     fetchresult : boolean
         Fetch the result of the query
     result : <varies>
@@ -46,20 +46,20 @@ class QueryJob(Swift_TOO_Status):
 
     # These are the kinds of results that can be returned from a QueryJob
     _subclasses = [
-        Swift_TOO_Status,
-        Swift_VisQuery,
-        Swift_AFST,
-        UVOT_mode,
-        Swift_PPST,
-        Swift_TOO_Requests,
-        Swift_Calendar,
-        Swift_GUANO,
-        Swift_Data,
-        Swift_Clock,
-        Swift_SAA,
+        TOOStatus,
+        VisQuery,
+        AFST,
+        UVOTMode,
+        PPST,
+        TOORequests,
+        Calendar,
+        GUANO,
+        Data,
+        Clock,
+        SAA,
     ]
     # API name
-    # This is really just a Swift_TOO_Status request with a twist
+    # This is really just a Swift_TOOStatus request with a twist
     api_name = "Swift_TOO_Status"
 
     def __init__(self, *args, **kwargs):
@@ -73,8 +73,8 @@ class QueryJob(Swift_TOO_Status):
         shared_secret : str
             shared secret for TOO API (default 'anonymous')
         """
-        # Call the Swift_TOO_Status constructor with fetchresult=True to that make this a QueryJob
-        Swift_TOO_Status.__init__(
+        # Call the TOOStatus constructor with fetchresult=True to that make this a QueryJob
+        TOOStatus.__init__(
             self, *args, fetchresult=True, api_name=self.api_name, **kwargs
         )